requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Authentication
pyjwt==2.8.0

# Object Storage (MinIO/S3)
boto3==1.34.34

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment variables
python-dotenv==1.0.0

# For production
gunicorn==21.2.0
